scripts/requirements.txt
# Python dependencies for MediaWiki import scripts

# MediaWiki API client
mwclient>=0.10.1

# HTTP client for AnythingLLM API
requests>=2.28.0

# Environment variable management
python-dotenv>=1.0.0
